+++ packages/react/src/components/Select/Select.tsx
@@ -20,7 +20,9 @@ export type SelectProps = SingleSelectProps | MultiSelectProps;
 
 export type SingleSelectProps = SelectPropsBase & {
   multiple?: false;
-  onChange?: SingleOnChangeEvent;
+  onBlur?: SingleSelectEvent;
+  onChange?: SingleSelectEvent;
+  onFocus?: SingleSelectEvent;
   options: SingleSelectOption[];
   value?: string;
 };
@@ -28,7 +30,9 @@ export type SingleSelectProps = SelectPropsBase & {
 export type MultiSelectProps = SelectPropsBase & {
   deleteButtonLabel?: string;
   multiple: true;
-  onChange?: MultipleOnChangeEvent;
+  onBlur?: MultiSelectEvent;
+  onChange?: MultiSelectEvent;
+  onFocus?: MultiSelectEvent;
   options: MultiSelectOption[];
   value?: string[];
 };
@@ -53,8 +57,8 @@ export type MultiSelectOption = SingleSelectOption & {
   deleteButtonLabel?: string;
 };
 
-export type SingleOnChangeEvent = (value: string) => void;
-export type MultipleOnChangeEvent = (value: string[]) => void;
+export type SingleSelectEvent = (value: string) => void;
+export type MultiSelectEvent = (value: string[]) => void;
 
 const eventListenerKeys = {
   ArrowUp: 'ArrowUp',
@@ -70,7 +74,9 @@ const Select = (props: SelectProps) => {
     inputId,
     label,
     multiple,
+    onBlur,
     onChange,
+    onFocus,
     options,
     searchLabel,
     value,
@@ -80,18 +86,31 @@ const Select = (props: SelectProps) => {
     throw Error('Each value in the option list must be unique.');
   }
 
+  const findOptionFromValue = useCallback(
+    (v?: string) =>
+      options.find((option) => option.value === v) ?? {
+        label: '',
+        value: '',
+      },
+    [options],
+  );
+
   // List of selected values if multiselect.
   const [selectedValues, setSelectedValues] = useState<string[]>(
     multiple ? value ?? [] : [],
   );
 
-  const [keyword, setKeyword] = useState('');
+  const [keyword, setKeyword] = useState(
+    !multiple ? findOptionFromValue(value)?.label ?? '' : '',
+  );
 
   const [sortedOptions, setSortedOptions] = useState(options);
   // Enable dynamic change of options by resetting sortedOptions
   const prevOptions = usePrevious([...options]);
+  const prevValue = usePrevious(value);
   useUpdate(() => {
     // Update not on changed reference but on changed values inside the object
+    let shouldSetValue = false;
     if (
       options.length !== prevOptions?.length ||
       options.some(
@@ -99,17 +118,25 @@ const Select = (props: SelectProps) => {
       )
     ) {
       setSortedOptions(options);
-      setSelectedValues(multiple ? value ?? [] : []);
+      shouldSetValue = true;
+    }
+
+    if (
+      (!multiple && value !== prevValue) ||
+      (multiple &&
+        (typeof prevValue === 'string' || !arraysEqual(value, prevValue))) ||
+      shouldSetValue
+    ) {
+      if (multiple) {
+        setSelectedValues(value ?? []);
+      } else {
+        setActiveOption(value);
+        setKeyword(findOptionFromValue(value)?.label ?? '');
+      }
     }
   });
 
   const numberOfOptions = options.length;
-
-  // When order of sorted options changes (due to change of search keyword), select first option.
-  const firstOptionValue = sortedOptions[0]?.value;
-  useUpdate(() => {
-    firstOptionValue !== undefined && setActiveOption(firstOptionValue);
-  }, [firstOptionValue]);
 
   // If multiselect, activeOption defines which option that has focus.
   // If single select, it defines the selected value.
@@ -126,24 +153,31 @@ const Select = (props: SelectProps) => {
     [setKeyword, multiple],
   );
 
-  const [usingKeyboard, setUsingKeyboard] = useState<boolean>(false);
-  useEventListener('click', () => setUsingKeyboard(false));
-  useEventListener('keydown', () => setUsingKeyboard(true));
-
-  const [expanded, setExpanded] = useState<boolean>(false);
-
   const listboxWrapperRef = useRef<HTMLSpanElement>(null);
   const selectFieldRef = useRef<HTMLSpanElement>(null);
 
-  useEffect(() => {
-    // Rerender when the value property changes
-    if (!multiple) {
-      setActiveOption(value);
-      setKeyword(findOptionFromValue(value)?.label ?? '');
-    } else if (!arraysEqual(value, selectedValues)) {
-      setSelectedValues(value ?? []);
+  const [usingKeyboard, setUsingKeyboard] = useState<boolean>(false);
+  const [hasFocus, setHasFocus] = useState<boolean>(false);
+  useEventListener('click', () => setUsingKeyboard(false));
+  useEventListener('keydown', () => setUsingKeyboard(true));
+  const updateHasFocus = () => {
+    const { activeElement } = document;
+    setHasFocus(selectFieldRef.current?.contains(activeElement) ?? false);
+  };
+  useEventListener('focusin', updateHasFocus);
+  useEventListener('focusout', updateHasFocus);
+
+  useUpdate(() => {
+    if (!multiple && !hasFocus) {
+      setKeyword(findOptionFromValue(activeOption)?.label ?? '');
     }
-  }, [value]);
+    if (hasFocus && onFocus)
+      multiple ? onFocus(selectedValues) : onFocus(activeOption || '');
+    else if (!hasFocus && onBlur)
+      multiple ? onBlur(selectedValues) : onBlur(activeOption || '');
+  }, [hasFocus]);
+
+  const [expanded, setExpanded] = useState<boolean>(false);
 
   useEffect(() => {
     // Ensure that active option is always visible when using keyboard
@@ -174,18 +208,12 @@ const Select = (props: SelectProps) => {
     }
   }, [activeOptionIndex]);
 
-  const findOptionFromValue = (v?: string) =>
-    options.find((option) => option.value === v) ?? {
-      label: '',
-      value: '',
-    };
-
   const multipleChangeHandler = (newValues: string[], addedValue?: string) => {
     if (!selectedValues?.length) {
       setActiveOption(addedValue);
     }
     setSelectedValues(newValues);
-    onChange && (onChange as MultipleOnChangeEvent)(newValues);
+    onChange && (onChange as MultiSelectEvent)(newValues);
     resetKeyword();
   };
 
@@ -193,7 +221,7 @@ const Select = (props: SelectProps) => {
     setActiveOption(newValue);
     resetKeyword(findOptionFromValue(newValue).label);
     setExpanded(false);
-    onChange && (onChange as SingleOnChangeEvent)(newValue);
+    onChange && (onChange as SingleSelectEvent)(newValue);
   };
 
   const addOrRemoveSelectedValue = (activeValue: string) => {
@@ -287,7 +315,16 @@ const Select = (props: SelectProps) => {
     const newKeyword = e.target.value;
     if (newKeyword) {
       // Update sorted options only if keyword has a non-empty value
-      setSortedOptions(optionSearch(options, newKeyword));
+      const newSortedOptions = optionSearch(options, newKeyword);
+      setSortedOptions(newSortedOptions);
+
+      // When order of sorted options changes (due to change of search keyword), select first option.
+      const firstOptionValue = sortedOptions[0]?.value;
+      const newFirstOptionValue = newSortedOptions[0]?.value;
+      if (newSortedOptions && firstOptionValue != newFirstOptionValue) {
+        setActiveOption(newFirstOptionValue);
+      }
+
       !expanded && setExpanded(true);
     }
     setKeyword(newKeyword);
@@ -342,7 +379,9 @@ const Select = (props: SelectProps) => {
                 </>
               )}
               <input
-                aria-activedescendant={`${id}-${activeOption}`}
+                aria-activedescendant={
+                  activeOption ? `${id}-${activeOption}` : undefined
+                }
                 aria-autocomplete='list'
                 aria-controls={listboxId}
                 aria-expanded={expanded}